eme_playback/test_eme_playback.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test Encrypted Media Extensions</title>
  <script type="text/javascript" src="SimpleTest.js"></script>
  <script type="text/javascript" src="manifest.js"></script>
  <script type="text/javascript" src="eme.js"></script>
</head>
<body>
<video name='v' id='video'>
<pre id="test">
<script class="testbody" type="text/javascript">

var EMEmanifest = [
  {
    name:"bipbop 10s",
    tracks: [
      {
          name:"video",
          type:"video/mp4; codecs=\"avc1.4d4015\"",
          fragments:[ "bipbop-cenc-video-10s.mp4",
                    ]
      }
    ],
    keys: {
      "7e571d037e571d037e571d037e571d11" : "7e5733337e5733337e5733337e573311",
    },
    sessionType:"temporary",
    sessionCount:1,
    duration:10.01
  },
];

function startTest(test, token)
{
  var sessions = [];
  function onSessionCreated(session) {
    sessions.push(session);
  }

  let v = document.getElementById("video");

  var gotEncrypted = 0;
  let finish = new EMEPromise;

  v.addEventListener("ended", function(ev) {
    ok(true, TimeStamp(token) + " got ended event");

  });

  Promise.all([
    LoadInitData(v, test, token),
    CreateAndSetMediaKeys(v, test, token),
    LoadTest(test, v, token)])
  .then(values => {
    console.log(" Start to play ENCRYPTED content !!");
    v.play();
    let initData = values[0];
    initData.map(ev => {
      let session = v.mediaKeys.createSession();
      onSessionCreated(session);
      MakeRequest(test, token, ev, session);
    });
  })
  .then(() => {
    console.log(" finish.promise ");
    return finish.promise;
  })
  .catch(reason => ok(false, reason))
}
startTest(EMEmanifest[0], 'test>>>>');
</script>
</pre>
</body>
</html>
